Update ESP crates, downgrade embassy crate
Updated to the ESP release candidate and removed like four dependencies. It's crazy that this tiny project is roughly 216 dependencies, though a significant portion are all `esp-` and `embassy-` crates depending on each other. Another portion is the fact that the "sh1106" driver is so outdated it requires an older version of the hal crate...