chore: Work on placing kerfur face elements
The mouth is not finished yet, not decided on how it will be formed.

The neutral face is mostly complete, but both meow and dazed are just starting.
Animations are somewhat fixed, but still don't properly return to their starting position.
Needed to add `libm` for access to square roots.